+++ src/main/java/antojos/ecommerce/products/ProductController.java
@@ -5,30 +5,29 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.*;
 
+import antojos.ecommerce.auth.Verifier;
 import antojos.ecommerce.products.drink.Drink;
 import antojos.ecommerce.products.drink.DrinkService;
 import antojos.ecommerce.products.food.Food;
 import antojos.ecommerce.products.food.FoodService;
+import antojos.ecommerce.user.Role;
+import antojos.ecommerce.user.User;
 import jakarta.servlet.http.HttpSession;
-import org.hibernate.Session;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+@AllArgsConstructor
 @Controller
 @RequestMapping({"/","/products"})
 public class ProductController {
   private ProductService productService;
   private FoodService foodService;
   private DrinkService drinkService;
+  private Verifier verifier;
 
-  public ProductController(ProductService productService, FoodService foodService, DrinkService drinkService) {
-    this.productService = productService;
-    this.foodService = foodService;
-    this.drinkService = drinkService;
-  }
 
 
   @GetMapping("")
@@ -56,16 +55,26 @@ public class ProductController {
   }
 
   @GetMapping("/searchAdmin")
-  public String searchProdsAdm(@RequestParam("query") String query, Model model){
-    Map<String, List<Product>> productsByType = productService.getProducts(query);
+  public String searchProdsAdm(@RequestParam("query") String query, Model model, HttpSession session){
+    String tokenInSession = (String) session.getAttribute("token");
+    String dniUserInSession = ((User) session.getAttribute("user")).getDni();
 
-    model.addAttribute("product", new Product());
-    model.addAttribute("foods", productsByType.getOrDefault("food", new ArrayList<>()));
-    model.addAttribute("drinks", productsByType.getOrDefault("drink", new ArrayList<>()));
+    if (verifier.verifyToken(tokenInSession,dniUserInSession)){
+      if (Objects.equals(verifier.verifyRole(session), Role.ADMIN)){
+        Map<String, List<Product>> productsByType = productService.getProducts(query);
 
-    return "/products/prodlist";
+        model.addAttribute("product", new Product());
+        model.addAttribute("foods", productsByType.getOrDefault("food", new ArrayList<>()));
+        model.addAttribute("drinks", productsByType.getOrDefault("drink", new ArrayList<>()));
+
+        return "/products/prodlist";
+      }else {
+        return "/users/login";
+      }
+    }else {
+      return "/users/login";
+    }
   }
-
 
 
 
@@ -155,16 +164,25 @@ public class ProductController {
 
   @GetMapping("/prodsList")
   public String productList(Model model, HttpSession session){
-    Map<String, List<Product>> productsByType = productService.getProducts("");
+    String tokenInSession = (String) session.getAttribute("token");
+    String dniUserInSession = ((User) session.getAttribute("user")).getDni();
 
-    model.addAttribute("product", new Product());
-    model.addAttribute("foods", productsByType.getOrDefault("food", new ArrayList<>()));
-    model.addAttribute("drinks", productsByType.getOrDefault("drink", new ArrayList<>()));
+    if (verifier.verifyToken(tokenInSession,dniUserInSession)){
+      if (Objects.equals(verifier.verifyRole(session), Role.ADMIN)){
+        Map<String, List<Product>> productsByType = productService.getProducts("");
 
+        model.addAttribute("product", new Product());
+        model.addAttribute("foods", productsByType.getOrDefault("food", new ArrayList<>()));
+        model.addAttribute("drinks", productsByType.getOrDefault("drink", new ArrayList<>()));
 //    resetErrorShowed(session);
 //    session.setAttribute("error_showed", false);
-
-    return "/products/prodlist";
+        return "/products/prodlist";
+      }else {
+        return "/users/login";
+      }
+    }else {
+      return "/users/login";
+    }
   }
 
 
@@ -187,47 +205,70 @@ public class ProductController {
   @PostMapping("/newProd")
   //, @RequestParam("imageProd") MultipartFile img
   public String createNewProd(@ModelAttribute("product") Product product, Model model, HttpSession session, @RequestParam("lts") float lts, @RequestParam("amountPeople") int amountPeople, @RequestParam("editType") String type, @RequestParam("imageProd") MultipartFile img){
-    boolean flag_prodExist = productService.verifyProdByName(product.getName());
+    String tokenInSession = (String) session.getAttribute("token");
+    String dniUserInSession = ((User) session.getAttribute("user")).getDni();
+
+    if (verifier.verifyToken(tokenInSession,dniUserInSession)){
+      if (Objects.equals(verifier.verifyRole(session), Role.ADMIN)){
+        boolean flag_prodExist = productService.verifyProdByName(product.getName());
 
 
-    if (!flag_prodExist){
-      if (Objects.equals(type, "food")){
-        Food food = new Food();
-        food.setAmountPeopleEat(amountPeople);
-        food.setStock(product.getStock());
-        food.setName(product.getName());
-        food.setDescription(product.getDescription());
-        food.setPrice(product.getPrice());
+        if (!flag_prodExist){
+          if (Objects.equals(type, "food")){
+            Food food = new Food();
+            food.setAmountPeopleEat(amountPeople);
+            food.setStock(product.getStock());
+            food.setName(product.getName());
+            food.setDescription(product.getDescription());
+            food.setPrice(product.getPrice());
 
-        foodService.addFood(food);
+            foodService.addFood(food);
 
-      }else if(Objects.equals(type, "drink")){
-        Drink drink = new Drink();
-        drink.setMililts(lts);
-        drink.setStock(product.getStock());
-        drink.setName(product.getName());
-        drink.setDescription(product.getDescription());
-        drink.setPrice(product.getPrice());
+          }else if(Objects.equals(type, "drink")){
+            Drink drink = new Drink();
+            drink.setLts(lts);
+            drink.setStock(product.getStock());
+            drink.setName(product.getName());
+            drink.setDescription(product.getDescription());
+            drink.setPrice(product.getPrice());
 
-        drinkService.addDrink(drink);
+            drinkService.addDrink(drink);
 
+          }
+          saveImage(product.getName(), img);
+        }else {
+          session.setAttribute("prodExist_error", "The product: " + product.getName() + " already exists ");
+          session.setAttribute("error_showed", true);
+        }
+
+        return "redirect:/products/prodsList";
+      }else {
+        return "/users/login";
       }
-      saveImage(product.getName(), img);
     }else {
-      session.setAttribute("prodExist_error", "The product: " + product.getName() + " already exists ");
-      session.setAttribute("error_showed", true);
+      return "/users/login";
     }
-
-    return "redirect:/products/prodsList";
   }
 
 
 
 
+  //MEPA Q NO SE USA
   @GetMapping("/delete/{id}")
-  public String deleteProd(@PathVariable Long id){
-    productService.deleteProduct(id);
-    return "redirect:/products/list";
+  public String deleteProd(@PathVariable Long id, HttpSession session){
+    String tokenInSession = (String) session.getAttribute("token");
+    String dniUserInSession = ((User) session.getAttribute("user")).getDni();
+
+    if (verifier.verifyToken(tokenInSession,dniUserInSession)){
+      if (Objects.equals(verifier.verifyRole(session), Role.ADMIN)){
+        productService.deleteProduct(id);
+        return "redirect:/products/list";
+      }else {
+        return "/users/login";
+      }
+    }else {
+      return "/users/login";
+    }
   }
