+++ src/main/java/antojos/ecommerce/products/drink/DrinkController.java
@@ -4,57 +4,77 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.Objects;
 
-import antojos.ecommerce.products.food.Food;
+import antojos.ecommerce.auth.Verifier;
+import antojos.ecommerce.user.Role;
+import antojos.ecommerce.user.User;
+import jakarta.servlet.http.HttpSession;
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-
+@AllArgsConstructor
 @Controller
 @RequestMapping("/drinks")
 public class DrinkController {
   private DrinkService drinkService;
-
-  public DrinkController(DrinkService drinkService) {
-    this.drinkService = drinkService;
-  }
+  private Verifier verifier;
 
 
   @PostMapping("/edit")
-  public String editDrink(@RequestParam Long id, @RequestParam String name, @RequestParam String desc, @RequestParam Float price, @RequestParam int stock, @RequestParam Float lts){
-    Drink drink = new Drink();
-    drink.setId(id);
-    drink.setName(name);
-    drink.setDescription(desc);
-    drink.setPrice(price);
-    drink.setStock(stock);
-    drink.setMililts(lts);
-    drinkService.updateDrink(drink);
-    return "redirect:/products/prodsList";
+  public String editDrink(@RequestParam Long id, @RequestParam String name, @RequestParam String desc, @RequestParam Float price, @RequestParam int stock, @RequestParam Float lts, HttpSession session){
+    String tokenInSession = (String) session.getAttribute("token");
+    String dniUserInSession = ((User) session.getAttribute("user")).getDni();
+
+    if (verifier.verifyToken(tokenInSession,dniUserInSession)){
+      if (Objects.equals(verifier.verifyRole(session), Role.ADMIN)){
+        Drink drink = new Drink();
+        drink.setId(id);
+        drink.setName(name);
+        drink.setDescription(desc);
+        drink.setPrice(price);
+        drink.setStock(stock);
+        drink.setLts(lts);
+        drinkService.updateDrink(drink);
+        return "redirect:/products/prodsList";
+      }else {
+        return "/users/login";
+      }
+    }else {
+      return "/users/login";
+    }
   }
 
 
 
-  private void deleteImgDrink(Long id){
+  private void deleteImgDrink(Long id) throws IOException{
     Drink drink = drinkService.getDrinkById(id);
 
     if(drink != null){
       String path = "src/main/resources/static/imgs/"+drink.getName()+".png";
-      try {
-        Files.deleteIfExists(Paths.get(path));
-      } catch (IOException e) {
-        e.printStackTrace();
-      }
+      Files.deleteIfExists(Paths.get(path));
     }
   }
 
 
   @PostMapping("delete")
-  public String deleteDrink(@RequestParam("id") Long id){
-    drinkService.deleteDrink(id);
-    deleteImgDrink(id);
-    return "redirect:/products/prodsList";
+  public String deleteDrink(@RequestParam("id") Long id, HttpSession session) throws IOException{
+    String tokenInSession = (String) session.getAttribute("token");
+    String dniUserInSession = ((User) session.getAttribute("user")).getDni();
+
+    if (verifier.verifyToken(tokenInSession,dniUserInSession)){
+      if (Objects.equals(verifier.verifyRole(session), Role.ADMIN)){
+        drinkService.deleteDrink(id);
+        deleteImgDrink(id);
+        return "redirect:/products/prodsList";
+      }else {
+        return "/users/login";
+      }
+    }else {
+      return "/users/login";
+    }
   }
